% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{calc_frip}
\alias{calc_frip}
\title{Calculate FRiP (Fraction of Reads in Peaks)}
\usage{
calc_frip(fragments, peaks)
}
\arguments{
\item{fragments}{GRanges object with fragment coordinates and cell barcode metadata}

\item{peaks}{GRanges object of peak regions (e.g., from MACS2 or ArchR)}
}
\value{
data.table with cell barcodes and FRiP values
}
\description{
Calculate FRiP (Fraction of Reads in Peaks)
}
