% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertions.R
\name{compute_car_coverage}
\alias{compute_car_coverage}
\title{Compute coverage of CAR insertion sites per cell}
\usage{
compute_car_coverage(fragments, car_sites)
}
\arguments{
\item{fragments}{GRanges object with cell barcode metadata}

\item{car_sites}{GRanges object of CAR insertion regions}
}
\value{
data.table with per-cell insertion site overlap counts
}
\description{
Compute coverage of CAR insertion sites per cell
}
