% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_utils.R
\name{load_car_sites}
\alias{load_car_sites}
\title{Load CAR-T insertion sites from BED file}
\usage{
load_car_sites(bed_path)
}
\arguments{
\item{bed_path}{Path to BED file}
}
\value{
GRanges object with insertion site ranges
}
\description{
Load CAR-T insertion sites from BED file
}
