% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_utils.R
\name{load_fragments}
\alias{load_fragments}
\title{Load fragment.tsv.gz file as GRanges}
\usage{
load_fragments(path, barcode_filter = NULL)
}
\arguments{
\item{path}{Path to the fragment.tsv.gz file}

\item{barcode_filter}{Optional vector of barcodes to keep}
}
\value{
GRanges object with barcode metadata column
}
\description{
Load fragment.tsv.gz file as GRanges
}
