% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_qc_report}
\alias{render_qc_report}
\title{Render CARTEpigenoQC HTML report}
\usage{
render_qc_report(input_rmd, output_html, params = list())
}
\arguments{
\item{input_rmd}{Path to the R Markdown report template}

\item{output_html}{Path to output HTML file}

\item{params}{List of named parameters to pass to the report}
}
\value{
Path to the rendered HTML file
}
\description{
Render CARTEpigenoQC HTML report
}
